core: Rename source files
