+++ src/core/electrostatics_magnetostatics/mmm-modpsi.cpp
@@ -22,6 +22,7 @@
 #include "config.hpp"
 
 #include "mmm-common.hpp"
+#include "mmm-modpsi.hpp"
 
 #include <cmath>
