+++ src/core/electrostatics_magnetostatics/mmm-modpsi.hpp
